Move kubectl plugins into github.com/kcp-dev/kcp/cli module
