+++ Dockerfile
@@ -30,6 +30,8 @@ COPY go.mod go.mod
 COPY go.sum go.sum
 COPY sdk/go.mod sdk/go.mod
 COPY sdk/go.sum sdk/go.sum
+COPY cli/go.mod cli/go.mod
+COPY cli/go.sum cli/go.sum
 USER 0
 
 # Install kubectl.
