+++ cli/pkg/bind/plugin/bind.go
@@ -31,8 +31,8 @@ import (
 	"k8s.io/cli-runtime/pkg/genericclioptions"
 	"k8s.io/client-go/rest"
 
-	"github.com/kcp-dev/kcp/pkg/cliplugins/base"
-	pluginhelpers "github.com/kcp-dev/kcp/pkg/cliplugins/helpers"
+	"github.com/kcp-dev/kcp/cli/pkg/base"
+	pluginhelpers "github.com/kcp-dev/kcp/cli/pkg/helpers"
 	apisv1alpha1 "github.com/kcp-dev/kcp/sdk/apis/apis/v1alpha1"
 	kcpclientset "github.com/kcp-dev/kcp/sdk/client/clientset/versioned/cluster"
 )
