+++ cli/pkg/crd/plugin/snapshot_test.go
