--- pkg/cmd/OWNERS
@@ -1,6 +1,0 @@
-approvers:
-- ncdc
-- sttts
-reviewers:
-- stevekuznetsov
-- davidfestal
